# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/conversations/{refId}": {
      "get": {
        "operationId": "InboxController_thread_v1",
        "parameters": [
          {
            "name": "refId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "One conversation and its messages",
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/v1/conversations/{refId}/reply": {
      "post": {
        "description": "Returns 202 with status \"pending\": the message and its outbound ledger row are written in one transaction and the platform call happens afterwards, in the relay. Send the same idempotencyKey to retry safely — a repeat returns the original message rather than a second one.",
        "operationId": "InboxController_reply_v1",
        "parameters": [
          {
            "name": "refId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Reply to a conversation",
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/v1/conversations/{refId}/assign": {
      "post": {
        "operationId": "InboxController_assign_v1",
        "parameters": [
          {
            "name": "refId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Assign a conversation to a team employee, or unassign it",
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/v1/conversations/{refId}/status": {
      "post": {
        "operationId": "InboxController_setStatus_v1",
        "parameters": [
          {
            "name": "refId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Move a conversation through its workflow",
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/v1/conversations/{refId}/resync": {
      "post": {
        "description": "Repairs a gap left by a webhook that never arrived. Meta serves only the 20 most recent messages of a thread, so this recovers a recent gap and cannot reach further back. Safe to call repeatedly: the projector dedups on the platform message id, and a resync already in flight is reported as queued: false rather than started twice.",
        "operationId": "InboxController_resync_v1",
        "parameters": [
          {
            "name": "refId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Re-read this thread from the platform",
        "tags": [
          "inbox"
        ]
      }
    },
    "/api/v1/conversations/{refId}/read": {
      "post": {
        "description": "An explicit write, rather than a side effect of reading the thread.",
        "operationId": "InboxController_markRead_v1",
        "parameters": [
          {
            "name": "refId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Clear the unread badge",
        "tags": [
          "inbox"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wouchh Unified API",
    "description": "Social inbox, comments and posts for connected enterprises. Every response uses one envelope: { success, data, meta } or { success: false, error: { code, message, details }, meta }.",
    "version": "dev",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
...
  }
}